model-training/requirements.txt
# Model Training Dependencies

# Core ML libraries
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
xgboost>=1.7.0
lightgbm>=3.3.0

# Chemistry libraries
rdkit-pypi>=2022.9.5

# Deep learning (optional)
torch>=2.0.0
transformers>=4.30.0

# Hyperparameter optimization
optuna>=3.0.0
ray[tune]>=2.0.0

# Data processing
scipy>=1.7.0
joblib>=1.1.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.12.0
plotly>=5.10.0

# Jupyter notebooks
jupyter>=1.0.0
ipykernel>=6.15.0
notebook>=6.4.12

# Progress bars
tqdm>=4.64.0

# Data download
requests>=2.28.0
wget>=3.2

# Utilities
python-dotenv>=0.20.0
pyyaml>=6.0
